+++ settings.php
@@ -24,6 +24,8 @@
 
 defined('MOODLE_INTERNAL') || die;
 
+use format_flexsections\constants;
+
 if ($ADMIN->fulltree) {
     $url = new moodle_url('/admin/course/resetindentation.php', ['format' => 'flexsections']);
     $link = html_writer::link($url, get_string('resetindentation', 'admin'));
@@ -39,4 +41,12 @@ if ($ADMIN->fulltree) {
     $settings->add(new admin_setting_configcheckbox('format_flexsections/showsection0titledefault',
         get_string('showsection0titledefault', 'format_flexsections'),
         get_string('showsection0titledefaultdesc', 'format_flexsections'), 0));
+    $options = [
+        constants::COURSEINDEX_FULL => get_string('courseindexfull', 'format_flexsections'),
+        constants::COURSEINDEX_SECTIONS => get_string('courseindexsections', 'format_flexsections'),
+        constants::COURSEINDEX_NONE => get_string('courseindexnone', 'format_flexsections'),
+    ];
+    $settings->add(new admin_setting_configselect('format_flexsections/courseindexdisplay',
+        get_string('courseindexdisplay', 'format_flexsections'),
+        get_string('courseindexdisplaydesc', 'format_flexsections'), 0, $options));
 }