+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## UNRELEASED
 ### Added
 - Setting to show header for the General section
+- Setting how to show the course index (sections and activities, only sections,
+  do not display)
 ### Fixed
 - Fixed a bug when drag&drop of activities was not possible if the target
   section is empty